Move generator to internal
