+++ internal/generator/filters.go
