+++ internal/generator/templates/component.tf
