+++ internal/generator/templates/component/aws_variables.tf
