+++ internal/generator/templates/component/azure_variables.tf
