+++ internal/generator/templates/component/commercetools_variables.tf
