+++ internal/generator/templates/services.tf
