+++ internal/generator/templates/services/amplience.tf
