+++ internal/generator/templates/services/aws/endpoints.tf
