+++ internal/generator/templates/services/aws/main.tf
