+++ internal/generator/templates/services/aws/url_locals.tf
