+++ internal/generator/templates/services/azure/app_service_plans.tf
