+++ internal/generator/templates/services/azure/frontdoor.tf
