+++ internal/generator/templates/services/azure/frontdoor_endpoint.tf
