+++ internal/generator/templates/services/azure/main.tf
