+++ internal/generator/templates/services/azure/url_locals.tf
