+++ internal/generator/templates/services/commercetools/commercetools_frontend.tf
