+++ internal/generator/templates/services/commercetools/stores.tf
