+++ internal/generator/templates/services/contentful.tf
