+++ internal/generator/utils.go
