+++ internal/generator/variables.go
