+++ internal/generator/writer.go
