+++ internal/generator/writer_test.go
